# Replace UUID validation RegEx with manual validation