When analyzing a callgrind profile of a full library rescan, I noticed
that there are a lot of CPU instructions coming from libpcre2, with the
call originating in the UUID validation code.

Previously, the regular expression used for UUID validation was parsed
from a string for every invocation of the validation function. This is
unnecessary. A library rescan will validate thousands of UUIDs (roughly
250k calls to `Uuid::isValid` with 200'000 instructions per call on my
laptop for a library rescan). By moving the code to a static class
member, I achieved a 15% speedup for a full library scan.

---

Library scan duration for five consecutive LibrePCB starts using the release build:

Before:

- 12'419 ms
- 12'239 ms
- 11'992 ms
- 12'377 ms
- 11'766 ms

Avg: 12'159 ms

After:

- 11'366 ms
- 9'975 ms
- 10'113 ms
- 10'149 ms
- 10'020 ms

Avg: 10'325 ms (-15%)