@ubruhin 2020-01-05
Nice catch!

But the fix could be better and simpler :wink: Creating static objects is often dangerous, especially for complex types. Since the order of how static objects are initialized is not well defined, very weird behavior could be introduced if there are dependencies between static objects. In addition, these objects are initialized very early (before `main()`) and thus important things like creating a `QApplication` instance (which is essential for many Qt objects), or in this case initializing the libpcre2, are not yet done.

Thus I'd avoid static objects whenever possible. Usually that's easy by wrapping such objects with a static method like this:

```cpp
class Foo {
  static const Bar& createBar() {
    static Bar bar;
    return bar;
  }
};
```

In that case, the object will be created once when it is accessed the first time, i.e. much later than static objects.

But in the `Uuid` class you probably don't even need such a wrapper method, just add `static const` to the definition of the `QRegularExpression` (`const` only to be sure the objects doesn't get modified during its lifetime), that should (hopefully) do the trick :wink:

Btw, there are several other classes which use `QRegularExpression` in a similar way, would you be interested to take a look at them too?