@ubruhin 2020-01-05
And one more issue since `Uuid` is used in multiple threads: https://stackoverflow.com/questions/25581696/qregularexpression-matching-thread-safety