@eduardosm 2020-01-05
Why not just write the matching code manually instead of using a regex library? It should be quite simple in this case.

Something like:
```
auto isLowerHex = [](QChar chr) {
    return (chr >= QChar('0') && chr <= QChar('9')) || (chr >= QChar('a') && chr <= QChar('f'));
};
if (str.length() != 36) return false;
if (!isLowerHex(str[0])) return false;
if (!isLowerHex(str[1])) return false;
...
if (str[8] != QChar('-')) return false;
...
```

I believe this would be faster than using a regular expression compiled at runtime.