@dbrgn 2020-01-05
> Creating static objects is often dangerous

Thanks for the explanation! I'm used to [lazy_static](https://docs.rs/lazy_static/1.4.0/lazy_static/) and already expected static initialization of complex objects to be potentially fishy :slightly_smiling_face: 

> Why not just write the matching code manually instead of using a regex library?

I thought about that too, but a regex basically does that as well, but in a very optimized way that *might* make optimal use of CPU caches and things like SIMD instructions.

In any case, we would need to benchmark it. Any performance optimization without a benchmark is just guessing.