@dbrgn 2020-01-05
Btw, a compile-time regular expression library [like ctre](https://www.compile-time.re/meeting-cpp-2019/slides/#/3/0/4) would be very nice, but it requires C++17 (Clang 5+, GCC 7+).